% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proj.verse-package.R
\docType{package}
\name{proj.verse}
\alias{proj.verse-package}
\alias{proj.verse}
\title{\code{\link{proj.verse}}}
\description{
Scientific project development and management.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/edwardlavender/proj.verse/}
  \item Report bugs at \url{https://github.com/edwardlavender/proj.verses/issues/}
}

}
\author{
\strong{Maintainer}: Edward Lavender \email{edward.lavender@eawag.ch} (\href{https://orcid.org/0000-0002-8040-7489}{ORCID})

}
